eval/requirements.txt
# Pin versions to avoid breaking changes - Agent Framework is in preview
agent-framework-core==1.0.0b260107
agent-framework-azure-ai==1.0.0b260107
azure-ai-agentserver-core==1.0.0b10
azure-ai-agentserver-agentframework==1.0.0b10
agent-dev-cli
debugpy
python-dotenv
